--- apm-agent-plugins/apm-error-logging-plugin/src/test/java/co/elastic/apm/agent/error/logging/AbstractLoggingInstrumentationTest.java
@@ -1,65 +1,0 @@
-package co.elastic.apm.agent.error.logging;
-
-import co.elastic.apm.agent.MockReporter;
-import co.elastic.apm.agent.bci.ElasticApmAgent;
-import co.elastic.apm.agent.configuration.SpyConfiguration;
-import co.elastic.apm.agent.error.logging.JavaUtilLoggingInstrumentation;
-import co.elastic.apm.agent.impl.ElasticApmTracer;
-import co.elastic.apm.agent.impl.ElasticApmTracerBuilder;
-import co.elastic.apm.agent.okhttp.OkHttp3ClientInstrumentation;
-import co.elastic.apm.agent.servlet.ServletInstrumentation;
-import net.bytebuddy.agent.ByteBuddyAgent;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-
-import java.util.Arrays;
-
-@RunWith(value = SpringJUnit4ClassRunner.class)
-@WebAppConfiguration
-@TestConfiguration
-public class AbstractLoggingInstrumentationTest {
-
-    protected static ElasticApmTracer tracer;
-    protected static MockReporter reporter;
-    protected MockMvc mockMvc;
-
-    @Autowired
-    WebApplicationContext wac;
-
-    @BeforeClass
-    @BeforeAll
-    public static void beforeAll() {
-        reporter = new MockReporter();
-        tracer = new ElasticApmTracerBuilder()
-            .configurationRegistry(SpyConfiguration.createSpyConfig())
-            .reporter(reporter)
-            .build();
-        ElasticApmAgent.initInstrumentation(tracer, ByteBuddyAgent.install(),
-            Arrays.asList(new ServletInstrumentation(tracer), new JavaUtilLoggingInstrumentation(), new Slf4jLoggingInstrumentation()));
-    }
-
-    @AfterClass
-    @AfterAll
-    public static void afterAll() {
-        ElasticApmAgent.reset();
-    }
-
-    @Before
-    @BeforeEach
-    public void setup() {
-        this.mockMvc = MockMvcBuilders.webAppContextSetup(wac).build();
-    }
-}
